+++ src/main/java/solutions/fairdata/openrefine/metadata/commands/StoreDataCommand.java
@@ -38,7 +38,11 @@ import solutions.fairdata.openrefine.metadata.commands.response.storage.StoreDat
 import solutions.fairdata.openrefine.metadata.dto.audit.EventSource;
 import solutions.fairdata.openrefine.metadata.dto.storage.ExportFormatDTO;
 import solutions.fairdata.openrefine.metadata.dto.storage.StorageDTO;
-import solutions.fairdata.openrefine.metadata.storage.*;
+import solutions.fairdata.openrefine.metadata.storage.FTPStorage;
+import solutions.fairdata.openrefine.metadata.storage.Storage;
+import solutions.fairdata.openrefine.metadata.storage.StorageRegistryUtil;
+import solutions.fairdata.openrefine.metadata.storage.TripleStoreHTTPStorage;
+import solutions.fairdata.openrefine.metadata.storage.VirtuosoStorage;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;